coord: Extract timeline logic from group commit
This commit simplifies the group commit logic by extracting all of the
timeline advancement logic out into a separate method. Advancing all
timelines isn't strictly required for correctness. Extracting the logic
out helps simplify the overall group commit logic.
